# src/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Express API",
    "description": "Express API Information",
    "version": "1.0.0",
    "contact": {
      "name": "Amazing Developer"
    },
    "servers": [
      "http://localhost:3000"
    ]
  },
  "paths": {
    "/benefits": {
      "get": {
        "summary": "Fetch benefits",
        "responses": {
          "200": {
            "description": "Returns a list of Benefits.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Benefit"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/employees/{id}": {
      "get": {
        "summary": "Fetch an employee by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The unique identifier of the employee.",
            "schema": {
              "type": "integer",
              "example": 123
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns a list of Employees.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Employee"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update employee details",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The unique identifier of the employee.",
            "schema": {
              "type": "integer",
              "example": 123
            }
          }
        ],
        "requestBody": {
          "description": "Partial employee details to update",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "firstName": {
                    "type": "string",
                    "description": "The first name of the employee",
                    "example": "John"
                  },
                  "lastName": {
                    "type": "string",
                    "description": "The last name of the employee",
                    "example": "Smith"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Returns a list of Employees.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Employee"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Benefit": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "Employee": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "date_of_birth": {
            "type": "string"
          },
          "secret": {
            "type": "string"
          }
        }
      }
    }
  },
  "tags": []
}